+++ sources/relay.h
@@ -14,10 +14,10 @@ typedef void        (*od_relay_on_read_t)(od_relay_t*, size_t size);
 
 struct od_relay
 {
-	int                   packet;
-	int                   packet_skip;
+	size_t                packet;
+	size_t                packet_skip;
 	machine_msg_t        *packet_full;
-	int                   packet_full_pos;
+	size_t                packet_full_pos;
 	machine_iov_t        *iov;
 	machine_cond_t       *base;
 	od_io_t              *src;
@@ -210,11 +210,11 @@ od_relay_process(od_relay_t *relay, int *progress, char *data, size_t size)
 		if (size < (int)sizeof(kiwi_header_t))
 			return OD_UNDEF;
 
-		int body;
+		size_t body;
 		body = kiwi_read_size(data, sizeof(kiwi_header_t));
 		body -= sizeof(uint32_t);
 
-		int total = sizeof(kiwi_header_t) + body;
+		size_t total = sizeof(kiwi_header_t) + body;
 		if (size >= total) {
 			*progress = total;
 			return od_relay_on_packet(relay, data, total);
@@ -240,7 +240,7 @@ od_relay_process(od_relay_t *relay, int *progress, char *data, size_t size)
 	}
 
 	/* chunk */
-	int to_parse = relay->packet;
+	size_t to_parse = relay->packet;
 	if (to_parse > size)
 		to_parse = size;
 	*progress = to_parse;
